backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.27.0
pyotp>=2.9.0
requests>=2.31.0
NorenRestApiPy>=0.0.25
selenium>=4.18.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
